Moves common db insertion funcs to utils
